First Recompile with node and Edge package Reorganized
