+++ include/CXXGraph/CXXGraph.hpp
@@ -2,14 +2,14 @@
 #define __CXXGRAPH_H__
 
 #include "CXXGraph/CXXGraphConfig.h"
-#include "CXXGraph/Edge/DirectedEdge.hpp"
-#include "CXXGraph/Edge/DirectedWeightedEdge.hpp"
-#include "CXXGraph/Edge/Edge.hpp"
-#include "CXXGraph/Edge/UndirectedEdge.hpp"
-#include "CXXGraph/Edge/UndirectedWeightedEdge.hpp"
-#include "CXXGraph/Edge/Weighted.hpp"
+#include "CXXGraph/Edge/DirectedEdge.h"
+#include "CXXGraph/Edge/DirectedWeightedEdge.h"
+#include "CXXGraph/Edge/Edge.h"
+#include "CXXGraph/Edge/UndirectedEdge.h"
+#include "CXXGraph/Edge/UndirectedWeightedEdge.h"
+#include "CXXGraph/Edge/Weighted.h"
 #include "CXXGraph/Graph/Graph.hpp"
-#include "CXXGraph/Node/Node.hpp"
+#include "CXXGraph/Node/Node.h"
 #include "CXXGraph/Partitioning/CoordinatedPartitionState.hpp"
 #include "CXXGraph/Partitioning/CoordinatedRecord.hpp"
 #include "CXXGraph/Partitioning/EBV.hpp"
