+++ include/CXXGraph/Edge/DirectedEdge_impl.hpp
@@ -17,56 +17,17 @@
 /***	 License: AGPL v3.0 ***/
 /***********************************************************/
 
-#ifndef __CXXGRAPH_DIRECTEDEDGE_H__
-#define __CXXGRAPH_DIRECTEDEDGE_H__
+#ifndef __CXXGRAPH_DIRECTEDEDGE_IMPL_H__
+#define __CXXGRAPH_DIRECTEDEDGE_IMPL_H__
 
 #pragma once
 
-#include "Edge.hpp"
+#include "DirectedEdge_decl.h"
 
 namespace CXXGraph {
-// Smart pointers alias
-template <typename T>
-using unique = std::unique_ptr<T>;
-template <typename T>
-using shared= std::shared_ptr<T>;
-
+  
 using std::make_unique;
 using std::make_shared;
-
-template <typename T>
-class UndirectedEdge;
-
-template <typename T>
-class DirectedEdge;
-// ostream operator
-template <typename T>
-std::ostream &operator<<(std::ostream &o, const DirectedEdge<T> &edge);
-template <typename T>
-class DirectedEdge : public Edge<T> {
- public:
-  DirectedEdge(const CXXGraph::id_t id, const Node<T> &node1,
-               const Node<T> &node2);
-  DirectedEdge(const CXXGraph::id_t id, shared<const Node<T>> node1,
-               shared<const Node<T>> node2);
-  DirectedEdge(const CXXGraph::id_t id,
-               const std::pair<const Node<T> *, const Node<T> *> &nodepair);
-  DirectedEdge(const CXXGraph::id_t id,
-               const std::pair<shared<const Node<T>>, shared<const Node<T>>> &nodepair);
-  DirectedEdge(const Edge<T> &edge);
-  virtual ~DirectedEdge() = default;
-  const Node<T> &getFrom() const;
-  const Node<T> &getTo() const;
-  const std::optional<bool> isDirected() const override;
-  const std::optional<bool> isWeighted() const override;
-  // operator
-  explicit operator UndirectedEdge<T>() const {
-    return UndirectedEdge<T>(Edge<T>::getId(), Edge<T>::getNodePair());
-  }
-
-  friend std::ostream &operator<< <>(std::ostream &os,
-                                     const DirectedEdge<T> &edge);
-};
 
 template <typename T>
 DirectedEdge<T>::DirectedEdge(const CXXGraph::id_t id, const Node<T> &node1,
@@ -122,4 +83,4 @@ std::ostream &operator<<(std::ostream &os, const DirectedEdge<T> &edge) {
 }
 }  // namespace CXXGraph
 
-#endif  // __CXXGRAPH_DIRECTEDEDGE_H__
+#endif  // __CXXGRAPH_DIRECTEDEDGE_IMPL_H__
