+++ include/CXXGraph/Edge/Edge_impl.hpp
@@ -17,17 +17,12 @@
 /***	 License: AGPL v3.0 ***/
 /***********************************************************/
 
-#ifndef __CXXGRAPH_EDGE_H__
-#define __CXXGRAPH_EDGE_H__
+#ifndef __CXXGRAPH_EDGE_IMPL_H__
+#define __CXXGRAPH_EDGE_IMPL_H__
 
 #pragma once
 
-#include <memory>
-#include <optional>
-#include <utility>
-
-#include "CXXGraph/Node/Node.hpp"
-#include "CXXGraph/Utility/id_t.hpp"
+#include "CXXGraph/Edge/Edge_decl.h"
 
 namespace CXXGraph {
 // Smart pointers alias
@@ -38,42 +33,6 @@ using shared = std::shared_ptr<T>;
 
 using std::make_unique;
 using std::make_shared;
-
-template <typename T>
-class Edge;
-// ostream operator
-template <typename T>
-std::ostream &operator<<(std::ostream &o, const Edge<T> &edge);
-template <typename T>
-class Edge {
- private:
-  CXXGraph::id_t id = 0;
-  std::pair<shared<const Node<T>>, shared<const Node<T>>> nodePair;
-
- public:
-  Edge(const CXXGraph::id_t id, const Node<T> &node1, const Node<T> &node2);
-  Edge(const CXXGraph::id_t id, shared<const Node<T>> node1, shared<const Node<T>> node2);
-  Edge(const CXXGraph::id_t id,
-       const std::pair<const Node<T> *, const Node<T> *> &nodepair);
-  Edge(const CXXGraph::id_t id,
-       const std::pair<shared<const Node<T>>, shared<const Node<T>>> &nodepair);
-  virtual ~Edge() = default;
-  void setFirstNode(shared<const Node<T>> node);
-  void setSecondNode(shared<const Node<T>> node);
-  const unsigned long long getId() const;
-  const std::pair<shared<const Node<T>>, shared<const Node<T>>> &getNodePair() const;
-  shared<const Node<T>> getOtherNode(shared<const Node<T>> node) const;
-  virtual const std::optional<bool> isDirected() const;
-  virtual const std::optional<bool> isWeighted() const;
-  // operator
-  virtual bool operator==(const Edge<T> &b) const;
-  bool operator<(const Edge<T> &b) const;
-  // operator DirectedEdge<T>() const { return DirectedEdge<T>(id, nodePair); }
-  // operator UndirectedEdge<T>() const { return UndirectedEdge<T>(id,
-  // nodePair); }
-
-  friend std::ostream &operator<< <>(std::ostream &os, const Edge<T> &edge);
-};
 
 template <typename T>
 Edge<T>::Edge(const CXXGraph::id_t id, const Node<T> &node1,
@@ -166,4 +125,4 @@ std::ostream &operator<<(std::ostream &os, const Edge<T> &edge) {
 }
 }  // namespace CXXGraph
 
-#endif  // __CXXGRAPH_EDGE_H__
+#endif  // __CXXGRAPH_EDGE_IMPL_H__
