+++ include/CXXGraph/Edge/UndirectedEdge.h
@@ -17,31 +17,11 @@
 /***	 License: AGPL v3.0 ***/
 /***********************************************************/
 
-#ifndef __CXXGRAPH_PARTITIONING_PARTITIONSTRATEGY_H__
-#define __CXXGRAPH_PARTITIONING_PARTITIONSTRATEGY_H__
+#ifndef __CXXGRAPH_UNDIRECTEDEDGE_H__
+#define __CXXGRAPH_UNDIRECTEDEDGE_H__
 
 #pragma once
 
-#include "CXXGraph/Edge/Edge.h"
-#include "PartitionState.hpp"
+#include "UndirectedEdge_impl.hpp"
 
-namespace CXXGraph {
-// Smart pointers alias
-template <typename T>
-using unique = std::unique_ptr<T>;
-template <typename T>
-using shared= std::shared_ptr<T>;
-
-using std::make_unique;
-using std::make_shared;
-
-namespace Partitioning {
-template <typename T>
-class PartitionStrategy {
- public:
-  virtual void performStep(shared<const Edge<T>> t, shared<PartitionState<T>> Sstate) = 0;
-};
-}  // namespace Partitioning
-}  // namespace CXXGraph
-
-#endif  // __CXXGRAPH_PARTITIONING_PARTITIONSTRATEGY_H__
+#endif  // __CXXGRAPH_UNDIRECTEDEDGE_H__
