+++ include/CXXGraph/Node/Node_decl.h
@@ -17,31 +17,42 @@
 /***	 License: AGPL v3.0 ***/
 /***********************************************************/
 
-#ifndef __CXXGRAPH_PARTITIONING_PARTITIONSTRATEGY_H__
-#define __CXXGRAPH_PARTITIONING_PARTITIONSTRATEGY_H__
+#ifndef __CXXGRAPH_NODE_DECL_H__
+#define __CXXGRAPH_NODE_DECL_H__
 
 #pragma once
+#include <iostream>
 
-#include "CXXGraph/Edge/Edge.h"
-#include "PartitionState.hpp"
+#include "CXXGraph/Utility/id_t.hpp"
 
 namespace CXXGraph {
-// Smart pointers alias
 template <typename T>
-using unique = std::unique_ptr<T>;
+class Node;
 template <typename T>
-using shared= std::shared_ptr<T>;
+std::ostream &operator<<(std::ostream &os, const Node<T> &node);
+template <typename T>
+class Node {
+ private:
+  CXXGraph::id_t id = 0;
+  std::string userId = "";
+  T data;
+  void setId(const std::string &);
 
-using std::make_unique;
-using std::make_shared;
-
-namespace Partitioning {
-template <typename T>
-class PartitionStrategy {
  public:
-  virtual void performStep(shared<const Edge<T>> t, shared<PartitionState<T>> Sstate) = 0;
+  Node(const std::string &, const T &data);
+  // Move constructor
+  Node(const std::string &, T &&data) noexcept;
+  ~Node() = default;
+  const CXXGraph::id_t &getId() const;
+  const std::string &getUserId() const;
+  const T &getData() const;
+  void setData(T &&new_data);
+  // operator
+  bool operator==(const Node<T> &b) const;
+  bool operator<(const Node<T> &b) const;
+  friend std::ostream &operator<< <>(std::ostream &os, const Node<T> &node);
 };
-}  // namespace Partitioning
+
 }  // namespace CXXGraph
 
-#endif  // __CXXGRAPH_PARTITIONING_PARTITIONSTRATEGY_H__
+#endif  // __CXXGRAPH_NODE_DECL_H__
