+++ include/CXXGraph/Node/Node_impl.hpp
@@ -17,46 +17,19 @@
 /***	 License: AGPL v3.0 ***/
 /***********************************************************/
 
-#ifndef __CXXGRAPH_NODE_H__
-#define __CXXGRAPH_NODE_H__
+#ifndef __CXXGRAPH_NODE_IMPL_H__
+#define __CXXGRAPH_NODE_IMPL_H__
 
-#pragma once
-
-#include "CXXGraph/Utility/id_t.hpp"
-
+#include "Node_decl.h"
 #include <iomanip>
-#include <iostream>
 
 namespace CXXGraph {
+
 template <typename T>
 class Node;
-template <typename T>
-std::ostream &operator<<(std::ostream &os, const Node<T> &node);
-template <typename T>
-class Node {
- private:
-  CXXGraph::id_t id = 0;
-  std::string userId = "";
-  T data;
-  void setId(const std::string &);
-
- public:
-  Node(const std::string &, const T& data);
-  // Move constructor
-  Node(const std::string &, T&& data) noexcept;
-  ~Node() = default;
-  const CXXGraph::id_t &getId() const;
-  const std::string &getUserId() const;
-  const T &getData() const;
-  void setData(T&& new_data);
-  // operator
-  bool operator==(const Node<T> &b) const;
-  bool operator<(const Node<T> &b) const;
-  friend std::ostream &operator<< <>(std::ostream &os, const Node<T> &node);
-};
 
 template <typename T>
-Node<T>::Node(const std::string& id, const T& data) {
+Node<T>::Node(const std::string &id, const T &data) {
   this->userId = id;
   // the userid is set as sha512 hash of the user provided id
   setId(id);
@@ -64,7 +37,7 @@ Node<T>::Node(const std::string& id, const T& data) {
 }
 
 template <typename T>
-Node<T>::Node(const std::string& id, T&& data) noexcept {
+Node<T>::Node(const std::string &id, T &&data) noexcept {
   this->userId = id;
   // the userid is set as sha512 hash of the user provided id
   setId(id);
@@ -73,27 +46,6 @@ Node<T>::Node(const std::string& id, T&& data) noexcept {
 
 template <typename T>
 void Node<T>::setId(const std::string &inpId) {
-  // const unsigned char* userId = reinterpret_cast<const unsigned char
-  // *>((*inpId).c_str() ); unsigned char obuf[64]; unsigned long long obuf[8];
-  // SHA512(userId, (*inpId).length(), reinterpret_cast<unsigned char*>(obuf));
-  /**
-  // Transform byte-array to string
-  std::stringstream shastr;
-  shastr << std::hex << std::setfill('0');
-  int i = 0;
-  //unsigned long can only store 8 bytes so we truncate the hash to 8 bytes
-  for (const auto &byte: obuf)
-  {
-          shastr << std::setw(2) << static_cast<int>(byte);
-          i++;
-          if (i==8) break;
-  }
-  auto idStr =  shastr.str();
-  // convert hex string to unsigned long long
-  std::istringstream iss(idStr);
-  iss >> std::hex >> this->id;
-
-  **/
   this->id = std::hash<std::string>{}(inpId);
 }
 
@@ -113,7 +65,7 @@ const T &Node<T>::getData() const {
 }
 
 template <typename T>
-void Node<T>::setData(T&& new_data) {
+void Node<T>::setData(T &&new_data) {
   this->data = std::move(new_data);
 }
 
@@ -136,6 +88,7 @@ std::ostream &operator<<(std::ostream &os, const Node<T> &node) {
      << "  Id:\t" << node.userId << "\n  Data:\t" << node.data << "\n}";
   return os;
 }
+
 }  // namespace CXXGraph
 
-#endif  // __CXXGRAPH_NODE_H__
+#endif  // __CXXGRAPH_NODE_IMPL_H__
