+++ include/CXXGraph/Partitioning/EdgeBalancedVertexCut.hpp
@@ -25,7 +25,7 @@
 
 #include <chrono>
 
-#include "CXXGraph/Edge/Edge.hpp"
+#include "CXXGraph/Edge/Edge.h"
 #include "PartitionStrategy.hpp"
 #include "CXXGraph/Partitioning/Utility/Globals.hpp"
