+++ include/CXXGraph/Partitioning/WeightBalancedLibra.hpp
@@ -24,7 +24,7 @@
 
 #include <chrono>
 
-#include "CXXGraph/Edge/Edge.hpp"
+#include "CXXGraph/Edge/Edge.h"
 #include "PartitionStrategy.hpp"
 #include "CXXGraph/Partitioning/Utility/Globals.hpp"
